Finished setting up permissions.

+++ app/Http/Controllers/StoreController.php
@@ -14,22 +14,37 @@ class StoreController extends Controller
 
     public function index()
     {
-        return Store::query()->get(['name', 'code']);
+        if (!auth()->user()->is_admin)
+            abort(401);
+        
+        if (auth()->user()->is_admin)
+            return Store::all();
+        else
+            return auth()->user()->stores()->get(['name', 'code']);
     }
 
     public function store(Request $request)
     {
+        if (!auth()->user()->is_admin)
+            abort(401);
+        
         $params = $request->json()->all();
         $store = Store::create($params);
     }
 
     public function show($storeCode)
     {
+        if (!auth()->user()->is_admin)
+            abort(401);
+        
         return Store::findByCode($storeCode);
     }
 
     public function update(Request $request)
     {
+        if (!auth()->user()->is_admin)
+            abort(401);
+        
         $params = $request->json()->all();
         $store = Store::find($params['id']);
         $store->fill($params);
@@ -38,6 +53,9 @@ class StoreController extends Controller
 
     public function destroy($storeCode)
     {
+        if (!auth()->user()->is_admin)
+            abort(401);
+            
         Store::findByCode($storeCode)->delete();
     }
 }
